Session 2: temperature fix, data optimisations, zigzag plan
- evaluate.py: add temperature_moves parameter to MCTSAgent to prevent
  deterministic game repetition (was causing all 100 games to be
  identical); default 10 half-moves at temperature=1.0
- data.py: uint8 tensor storage, label_class tolerance function,
  zst support, numpy pre-allocation, improved validation checks
- train.py: add .float() conversion for uint8→float32 tensors
- ZIGZAG.md: full zigzag training curriculum design document
  (self-play → SF re-label → train loop, rounds 1-4)
- Remove src/__pycache__ from git tracking (.gitignore already covers it)

+++ src/train.py
@@ -69,7 +69,7 @@ def run_epoch(model, loader, optimizer, is_train: bool):
     ctx = torch.enable_grad() if is_train else torch.no_grad()
     with ctx:
         for tensors, values, move_idxs in loader:
-            tensors   = tensors.to(device)
+            tensors   = tensors.float().to(device)   # uint8 → float32
             values    = values.to(device)
             move_idxs = move_idxs.to(device)
